-NOT WORKING- halfway done fixing this shit
