--- src/main/java/com/magtable/model/Assignment.java
@@ -1,36 +1,0 @@
-package com.magtable.model;
-
-
-import javax.persistence.*;
-import java.io.Serializable;
-
-@Entity
-@Table(name = "assignment")
-public class Assignment implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "assignmentid", insertable = false, nullable = false)
-    private Integer assignmentID;
-
-    @Column(name = "magid", nullable = false)
-    private Integer magID;
-
-    public Integer getAssignmentID() {
-        return assignmentID;
-    }
-
-    public void setAssignmentID(Integer assignmentID) {
-        this.assignmentID = assignmentID;
-    }
-
-    public Integer getMagID() {
-        return magID;
-    }
-
-    public void setMagID(Integer magID) {
-        this.magID = magID;
-    }
-
-}
