+++ src/main/java/com/magtable/repository/AssignmentRepository.java
@@ -1,6 +1,5 @@
 package com.magtable.repository;
 
-import com.magtable.model.Assignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
