+++ src/main/java/com/magtable/repository/MagtablerecordRepository.java
@@ -1,6 +1,5 @@
 package com.magtable.repository;
 
-import com.magtable.model.Magtablerecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
