+++ src/main/java/com/magtable/repository/ParkinglocationRepository.java
@@ -1,6 +1,5 @@
 package com.magtable.repository;
 
-import com.magtable.model.Parkinglocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
