+++ src/main/java/com/magtable/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.magtable.repository;
 
-import com.magtable.model.Role;
+import com.magtable.model.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
