+++ src/main/java/com/magtable/repository/ShiftRepository.java
@@ -1,7 +1,6 @@
 package com.magtable.repository;
 
-import com.magtable.model.Shift;
-import com.magtable.model.W2WShift;
+import com.magtable.model.entities.Shift;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
