+++ src/main/java/com/magtable/repository/TruckRepository.java
@@ -1,6 +1,5 @@
 package com.magtable.repository;
 
-import com.magtable.model.Truck;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
