+++ src/main/java/com/magtable/repository/W2WShiftRepository.java
@@ -1,7 +1,6 @@
 package com.magtable.repository;
 
-import com.magtable.model.W2WShift;
-import org.springframework.data.domain.Sort;
+import com.magtable.model.entities.W2WShift;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
