+++ src/main/resources/database/MagTableDev.sql
@@ -26,97 +26,75 @@ CREATE TABLE User
     PRIMARY KEY (userID),
     CONSTRAINT FK_User_Role FOREIGN KEY (`role`) REFERENCES `Role` (`roleID`) ON DELETE RESTRICT ON UPDATE RESTRICT
 );
-
-CREATE TABLE Truck
-(
-    truckID INT(5) NOT NULL,
-    status  VARCHAR(4),
-    notice  VARCHAR(2000),
-    PRIMARY KEY (TruckID),
-    CONSTRAINT CK_Truck_Status CHECK (status = 'GO' OR status = 'INOP' OR status = 'CON' OR status = 'OOS')
-);
-
-CREATE TABLE BrixRecord
-(
-    brixRecordID    INT(5) NOT NULL AUTO_INCREMENT,
-    truckID         INT(5) NOT NULL,
-    nozzle          FLOAT(3),
-    type1           FLOAT(3),
-    type4           FLOAT(3),
-    litersPurged    INT(5),
-    timeMeasured    DATETIME,
-    PRIMARY KEY (BrixRecordID),
-    CONSTRAINT FK_BrixRecord_Assignment FOREIGN KEY (truckID) REFERENCES Truck (truckID) ON DELETE RESTRICT ON UPDATE RESTRICT,
-    CONSTRAINT CK_litersPurged CHECK (litersPurged >= 0 AND litersPurged <= 1000)
-);
-
-CREATE TABLE Tower
-(
-    towerID  INT(5)      NOT NULL,
-    position VARCHAR(25) NOT NULL,
-    PRIMARY KEY (towerID)
-);
-
-CREATE TABLE ParkingLocation
-(
-    id        INT(5)     NOT NULL,
-    apron     VARCHAR(4) NOT NULL,
-    phonetic  CHAR(1)    NOT NULL,
-    east      BOOLEAN    NOT NULL,
-    center    BOOLEAN    NOT NULL,
-    west      BOOLEAN    NOT NULL,
-    `left`    INT(2),
-    `right`   INT(2),
-    composite INT(2),
-    `double`  BOOLEAN,
-    PRIMARY KEY (id)
-);
-
 CREATE TABLE MagTableRecord
 (
-    magID         INT(5) NOT NULL AUTO_INCREMENT,
-    dailyMix      INT(2),
-    forecastLow   INT(2),
-    publishedBY   VARCHAR(32),
-    timePublished DATETIME,
-    PRIMARY KEY (magID)
+    magtableRecordID    INT(5) NOT NULL AUTO_INCREMENT,
+    dailyMix            INT(2),
+    forecastLow         INT(2),
+    publishedBy         VARCHAR(32),
+    timePublished       DATETIME,
+    PRIMARY KEY (magtableRecordID)
 );
 
 CREATE TABLE Assignment
 (
-    assignmentID    INT(10) NOT NULL AUTO_INCREMENT,
-    magID           INT(5)  NOT NULL,
+    assignmentID        INT(10) NOT NULL AUTO_INCREMENT,
+    magtableRecordID    INT(5)  NOT NULL,
     PRIMARY KEY (assignmentID),
-    CONSTRAINT FK_Mag_ID FOREIGN KEY (magID) REFERENCES MagTableRecord (magID) ON DELETE RESTRICT ON UPDATE RESTRICT
+    CONSTRAINT FK_Mag_ID FOREIGN KEY (magtableRecordID) REFERENCES MagTableRecord (magtableRecordID) ON DELETE RESTRICT ON UPDATE RESTRICT
 );
-
-CREATE TABLE PlAssignment
-(
-    plAssignmentID INT(10) NOT NULL AUTO_INCREMENT,
-    assignmentID   INT(10) NOT NULL,
-    id             INT(5),
-    position       VARCHAR(6),
-    bay            INT(2),
-    PRIMARY KEY (plAssignmentID),
-    CONSTRAINT FK_PLA_Assignment FOREIGN KEY (assignmentID) REFERENCES Assignment (assignmentID) ON DELETE RESTRICT ON UPDATE RESTRICT,
-    CONSTRAINT FK_PLA_PL FOREIGN KEY (id) REFERENCES ParkingLocation (id) ON DELETE RESTRICT ON UPDATE RESTRICT
-);
-
 
 CREATE TABLE Equipment
 (
-    equipmentID  INT(10) NOT NULL AUTO_INCREMENT,
-    id           INT(5)  NOT NULL,
+    equipmentID     INT(5)          NOT NULL,
+    type            VARCHAR(25)     NOT NULL,   -- tower description / truck type
+    status          VARCHAR(4),     -- truck operational status
+    notice          VARCHAR(2000),  -- truck operational notices
+    active          BOOLEAN,
+    PRIMARY KEY (equipmentID)
+);
+
+CREATE TABLE BrixRecord
+(
+    brixRecordID        INT(5) NOT NULL AUTO_INCREMENT,
+    equipmentID         INT(5) NOT NULL,
+    nozzle              FLOAT(3),
+    type1               FLOAT(3),
+    type4               FLOAT(3),
+    litersPurged        INT(5),
+    timeMeasured        DATETIME,
+    PRIMARY KEY (BrixRecordID),
+    CONSTRAINT FK_BrixRecord_Assignment FOREIGN KEY (equipmentID) REFERENCES Equipment (equipmentID) ON DELETE RESTRICT ON UPDATE RESTRICT,
+    CONSTRAINT CK_equipment_truck CHECK (equipmentID >= 0 AND equipmentID <= 1000), -- must be a truck
+    CONSTRAINT CK_litersPurged CHECK (litersPurged >= 0 AND litersPurged <= 1000)
+);
+
+CREATE TABLE AssignmentEquipment
+(
+    assignmentEquipmentID  INT(10) NOT NULL AUTO_INCREMENT,
     assignmentID INT(10) NOT NULL,
+    equipmentID  INT(5)  NOT NULL,
     status       VARCHAR(4),
     notice       VARCHAR(2000),
-    PRIMARY KEY (equipmentID),
-    CONSTRAINT FK_Equipment_Assignment FOREIGN KEY (assignmentID) REFERENCES Assignment (assignmentID) ON DELETE RESTRICT ON UPDATE RESTRICT
+    PRIMARY KEY (assignmentEquipmentID),
+    CONSTRAINT FK_Equipment_Assignment FOREIGN KEY (assignmentID) REFERENCES Assignment (assignmentID) ON DELETE RESTRICT ON UPDATE RESTRICT,
+    CONSTRAINT FK_Truck FOREIGN KEY (equipmentID) REFERENCES Equipment (equipmentID) ON DELETE RESTRICT ON UPDATE RESTRICT
+);
+
+CREATE TABLE AssignmentParkingLocation
+(
+    assignmentParkingLocationID INT(10) NOT NULL AUTO_INCREMENT,
+    assignmentID                INT(10) NOT NULL,
+    parkingLocationID           INT(5),
+    position                    VARCHAR(6),
+    bay                         INT(2),
+    PRIMARY KEY (assignmentParkingLocationID),
+    CONSTRAINT FK_PLA_Assignment FOREIGN KEY (assignmentID) REFERENCES Assignment (assignmentID) ON DELETE RESTRICT ON UPDATE RESTRICT
 );
 
 CREATE TABLE Shift
 (
-    shiftID      INT(10) NOT NULL AUTO_INCREMENT, #TODO Remove Auto Increment ( Inerts break with it)
+    shiftID      INT(10) NOT NULL,
     assignmentID INT(10) NOT NULL,
     description  VARCHAR(30),
     name         VARCHAR(50),
